third_party/blink/web_tests/external/wpt/shadow-dom/focus/focus-method-delegatesFocus-expected.txt

This is a testharness.js-based test.
PASS focus() on host with delegatesFocus, all tabindex=0
PASS focus() on host with delegatesFocus & tabindex =-1, all other tabindex=0
PASS focus() on host with delegatesFocus & no tabindex, all other tabindex=0
FAIL focus() on host with delegatesFocus & tabindex = 0, all other tabindex=-1 assert_equals: expected Element node <div tabindex="-1">aboveSlots</div> but got null
PASS focus() on host with delegatesFocus, all without tabindex
FAIL focus() on host with delegatesFocus, all tabindex=-1 assert_equals: expected Element node <div tabindex="-1">aboveSlots</div> but got null
PASS focus() on host with delegatesFocus & tabindex=0, #belowSlots with tabindex=0
PASS focus() on host with delegatesFocus & tabindex=0, #outside with tabindex=0
PASS focus() on host with delegatesFocus & tabindex=0, #aboveSlots and #belowSlots with tabindex=0
FAIL focus() on host with delegatesFocus & tabindex=0, #aboveSlots with tabindex=0 and #belowSlots with tabindex=1 assert_equals: expected Element node <div tabindex="0">aboveSlots</div> but got Element node <div tabindex="1">belowSlots</div>
PASS focus() on host with delegatesFocus & tabindex=0, #slottedToFirstSlot, #slottedToSecondSlot, #belowSlots  with tabindex=0
PASS focus() on host with delegatesFocus and already-focused non-first shadow descendant
PASS focus() on host with delegatesFocus with another host with no delegatesFocus and a focusable child
PASS focus() on host with delegatesFocus with another host with delegatesFocus and a focusable child
Harness: the test ran to completion.
